# openapi.yaml
openapi: 3.0.3
info:
  title: NutriCoach API
  description: |
    NutriCoach is an AI-powered nutrition tracking application. This API provides endpoints for food logging, 
    AI coaching, nutrition analysis, and progress tracking.
    
    ## Authentication
    All endpoints except `/api/healthz` require user authentication via session cookies.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Limits vary by endpoint type.
    
    ## Data Formats
    - All timestamps are in ISO 8601 format
    - Weights are in kilograms
    - Volumes are in milliliters
    - Nutrition values are per serving unless specified otherwise
  version: 1.0.0
  contact:
    name: NutriCoach Support
    url: https://github.com/your-repo/nutricoach
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server (Docker host port 5001 -> container 5000)
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /logs:
    get:
      summary: Get food logs
      description: Retrieve food logs for the authenticated user
      tags:
        - Food Logging
      parameters:
        - name: date
          in: query
          description: Filter logs by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Maximum number of logs to return
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of food logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/FoodLog'
                  count:
                    type: integer

    post:
      summary: Create food log entry
      description: Log a new food item
      tags:
        - Food Logging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFoodLog'
      responses:
        '201':
          description: Food log created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string
...
components:
  schemas:
    FoodLog:
      type: object
      properties:
        id:
          type: integer
        food_name:
          type: string
        meal:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        grams:
          type: number
          format: float
        calories:
          type: number
          format: float
        protein_g:
          type: number
          format: float
        carbs_g:
          type: number
          format: float
        fat_g:
          type: number
          format: float
        fiber_g:
          type: number
          format: float
        sugar_g:
          type: number
          format: float
        sodium_mg:
          type: number
          format: float
        source:
          type: string
          enum: [manual, vision, barcode]
        logged_at:
          type: string
          format: date-time
    CreateFoodLog:
      type: object
      required:
        - custom_name
        - meal
        - grams
        - calories
      properties:
        custom_name:
          type: string
        meal:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        grams:
          type: number
          format: float
        calories:
          type: number
          format: float
        protein_g:
          type: number
          format: float
          default: 0
        carbs_g:
          type: number
          format: float
          default: 0
        fat_g:
          type: number
          format: float
          default: 0
        fiber_g:
          type: number
          format: float
          default: 0
        sugar_g:
          type: number
          format: float
          default: 0
        sodium_mg:
          type: number
          format: float
          default: 0
        source:
          type: string
          enum: [manual, vision, barcode]
          default: manual
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
security:
  - sessionAuth: []
...